LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-transition-transform-function-none.tentative-expected.txt


FAIL A custom property keyword none is not a <transform-function> value and is not interpolable with <transform-function> values assert_equals: A single animation is running expected 1 but got 0
